+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+### Added
+
+-   All type declarations to the return type of requests.
+-   `PlaylistVersion` to encapsulate playlist modifications.
+
+## [0.6.0] - 2020-03-30
 
 ### Added